refactor(memify): Split consolidate_entity_descriptions into task modules (SDK-435)
The pipeline file had grown to 606 lines with 4 Pydantic models and ~29
functions - graph reads, prompt building, LLM calls, grouping, and
persistence-shape mutation all in one place, with the actual pipeline
wiring reduced to 13 lines at the very bottom. Every sibling in
memify_pipelines/ is a short orchestrator (33-116 lines) with the real
work living in cognee/tasks/memify/; this file broke that convention,
so its helpers couldn't be registered in memify_task_registry.py or
reused the way sibling memify tasks are.

Move the logic into cognee/tasks/memify/consolidate_entity_descriptions/,
split along the module's own pre-existing `# region` boundaries:
read_neighborhood.py (graph read + format_connections), rewrite_entities.py
(Entity rewrite stage), describe_types.py (EntityType stage +
apply_type_description), and models.py.

memify_pipelines/consolidate_entity_descriptions.py now keeps only the
Task list, and gains the dataset/user resolution it was missing -
mirroring consolidate_entities_pipeline's pattern (get_authorized_existing_datasets
+ set_database_global_context_variables) so it can target a dataset
other than main_dataset and run under a specific user, instead of
always falling through to memify()'s bare defaults.
