+++ cognee/tasks/memify/consolidate_entity_descriptions/__init__.py
@@ -1,0 +1,9 @@
+from .read_neighborhood import get_entities_with_neighborhood
+from .rewrite_entities import generate_consolidated_entities
+from .describe_types import generate_type_descriptions
+
+__all__ = [
+    "get_entities_with_neighborhood",
+    "generate_consolidated_entities",
+    "generate_type_descriptions",
+]
